Introduce a separate phase for shadow stack allocation
For now, just move the code from lowering to it.

Fixes the GC hole with CID resolvers.
