+++ src/coreclr/jit/CMakeLists.txt
@@ -228,6 +228,7 @@ set( JIT_WASM64_SOURCES
   llvm.cpp
   llvmtypes.cpp
   llvmlower.cpp
+  llvmlssa.cpp
   llvmcodegen.cpp
   llvmdebuginfo.cpp
 )
@@ -236,6 +237,7 @@ set( JIT_WASM32_SOURCES
   llvm.cpp
   llvmtypes.cpp
   llvmlower.cpp
+  llvmlssa.cpp
   llvmcodegen.cpp
   llvmdebuginfo.cpp
 )
